+++ example/gpu/bfs/bfs.cu
@@ -184,7 +184,7 @@ __global__ void bfs(uint32_t paramc, uint64_t * paramv, uint32_t depc, artsEdtDe
 void doPartionSync(uint32_t paramc, uint64_t * paramv, uint32_t depc, artsEdtDep_t depv[])
 {
     DPRINTF("Just Synced Partitions! %lu\n", paramv[0]);
-    artsSignalEdt(paramv[1], -1, NULL_GUID);
+    artsSignalEdt(paramv[1], (uint32_t)-1, NULL_GUID);
 }
 
 //There is only one of these per level.  It is signaled by the epoch containing the Bfs'es
@@ -209,7 +209,7 @@ void launchSort(uint32_t paramc, uint64_t * paramv, uint32_t depc, artsEdtDep_t 
     //We need the nextBfsEpoch and the nextLaunchBfsGuids to kick off launchBfs...
     dim3 threads(1, 1, 1);
     dim3 grid(1, 1, 1);
-    uint64_t args[] = {localLevel, nextLaunchBfsGuid};
+    uint64_t args[] = {localLevel, (uint64_t)nextLaunchBfsGuid};
     for(unsigned int j=0; j<artsGetTotalNodes(); j++)
     {
         for(uint64_t i=0; i<artsGetTotalGpus(); i++)
@@ -221,15 +221,15 @@ void launchSort(uint32_t paramc, uint64_t * paramv, uint32_t depc, artsEdtDep_t 
     {
         for(unsigned int i=0; i<artsGetTotalNodes(); i++)
         {
-            uint64_t syncArgs[] = {localLevel, nextLaunchBfsGuid};
-            artsGuid_t edtGuid = artsEdtCreate(doPartionSync, i, 2, syncArgs, partCount[i]);
+          uint64_t syncArgs[] = {localLevel, (uint64_t)nextLaunchBfsGuid};
+          artsGuid_t edtGuid =
+              artsEdtCreate(doPartionSync, i, 2, syncArgs, partCount[i]);
             DPRINTF("edtGuid: %lu\n", edtGuid);
             unsigned int slot = 0;
-            for(unsigned int j=0; j<PARTS; j++)
-            {
-                if(i == artsGuidGetRank(visitedGuid[j]))
-                {
-                    DPRINTF("Signaling: %lu with part: %lu\n", edtGuid, visitedGuid[j]);
+          for (unsigned int j = 0; j < PARTS; j++) {
+            if (i == artsGuidGetRank(visitedGuid[j])) {
+              DPRINTF("Signaling: %lu with part: %lu\n", edtGuid,
+                      visitedGuid[j]);
                     artsLCSync(edtGuid, slot++, visitedGuid[j]);
                 }
             }
